Isolate the visible part of an application inside a public folder.
index.php is symlinked in the public folder.

install.php was included in a relatively to the index.php location,
which is not right if the application is accessed form a symlink of
index.php. The inclusion is now relative to DOCROOT.

This is essential when you have to hide files like composer.json or
phinx.yml containing sensitive information.